[Fix] : 검색 결과 코드 분리
